# forge-go/api/openapi/codegen-openapi.json
{
  "components": {
    "schemas": {
      "AddMessageToBoardRequest": {
        "properties": {
          "message_id": {
            "title": "Message Id",
            "type": "string"
          }
        },
        "required": [
          "message_id"
        ],
        "title": "AddMessageToBoardRequest",
        "type": "object"
      },
...
      "BoardMessagesResponse": {
        "properties": {
          "ids": {
            "items": {
              "type": "string"
            },
            "title": "Ids",
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "title": "BoardMessagesResponse",
        "type": "object"
      },
      "BoardResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "guild_id": {
            "title": "Guild Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_default": {
            "title": "Is Default",
            "type": "boolean"
          },
          "is_private": {
            "title": "Is Private",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "guild_id",
          "created_by",
          "created_at",
          "is_default",
          "is_private"
        ],
        "title": "BoardResponse",
        "type": "object"
      },
      "BoardsResponse": {
        "properties": {
          "boards": {
            "items": {
              "$ref": "#/components/schemas/BoardResponse"
            },
            "title": "Boards",
            "type": "array"
          }
        },
        "required": [
          "boards"
        ],
        "title": "BoardsResponse",
        "type": "object"
      },
...
      "CreateBoardRequest": {
        "properties": {
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "guild_id": {
            "title": "Guild Id",
            "type": "string"
          },
          "is_default": {
            "default": false,
            "nullable": true,
            "title": "Is Default",
            "type": "boolean"
          },
          "is_private": {
            "default": false,
            "nullable": true,
            "title": "Is Private",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "guild_id",
          "name",
          "created_by"
        ],
        "title": "CreateBoardRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IdInfo": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "IdInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Rustic AI API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/addons/boards/": {
      "get": {
        "description": "Retrieves all boards for a guild.\n\nArgs:\n    guild_id (str): The ID of the guild to get boards for.\n\nReturns:\n    BoardsResponse: The boards for the guild.",
        "operationId": "getBoards",
        "parameters": [
          {
            "in": "query",
            "name": "guild_id",
            "required": true,
            "schema": {
              "title": "Guild Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Boards",
        "tags": [
          "boards"
        ]
      },
      "post": {
        "description": "Creates a new board for a guild.\n\nArgs:\n    guild_id (str): The ID of the guild to create the board for.\n    board (CreateBoardRequest): The board creation request.\n\nReturns:\n    IdInfo: The id of the newly created board.",
        "operationId": "createBoard",
        "parameters": [
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBoardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Board",
        "tags": [
          "boards"
        ]
      }
    },
    "/addons/boards/{board_id}/messages": {
      "get": {
        "description": "Retrieves all message IDs for a board.\n\nArgs:\n    board_id (str): The ID of the board to get messages for.\n\nReturns:\n    BoardMessagesResponse: The message IDs in the board.",
        "operationId": "getBoardMessageIds",
        "parameters": [
          {
            "in": "path",
            "name": "board_id",
            "required": true,
            "schema": {
              "title": "Board Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardMessagesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Board Messages",
        "tags": [
          "boards"
        ]
      },
      "post": {
        "description": "Adds a message to a board.\n\nArgs:\n    board_id (str): The ID of the board to add the message to.\n    request (AddMessageToBoardRequest): The request containing the message_id.\n\nReturns:\n    dict: Success message.",
        "operationId": "addMessageToBoard",
        "parameters": [
          {
            "in": "path",
            "name": "board_id",
            "required": true,
            "schema": {
              "title": "Board Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMessageToBoardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Message To Board",
        "tags": [
          "boards"
        ]
      }
    },
    "/addons/boards/{board_id}/messages/{message_id}": {
      "delete": {
        "description": "Removes a message from a board.\n\nArgs:\n    board_id (str): The ID of the board to remove the message from.\n    message_id (str): The ID of the message to remove.\n\nReturns:\n    dict: Success message.",
        "operationId": "removeMessageFromBoard",
        "parameters": [
          {
            "in": "path",
            "name": "board_id",
            "required": true,
            "schema": {
              "title": "Board Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "title": "Message Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sqldb",
            "required": false,
            "schema": {
              "default": "sqlite:///rustic_app.db",
              "title": "Sqldb",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Message From Board",
        "tags": [
          "boards"
        ]
      }
    },
...
  }
}